Move ajax -> service folder
All of our service streams will go here. Right now, this only contains
the ajax$ service, but other side-effect-producing services can
be contained here as well. If they're good, we can migrate them
back into brookjs.
