+++ client/delta/search.js
@@ -2,7 +2,7 @@
 import type { Observable } from 'kefir';
 import type { Action, TinyMCEState } from '../type';
 import R from 'ramda';
-import ajax from '../ajax';
+import { ajax$ } from '../service';
 import { SEARCH_INPUT, searchResultsSucceededAction } from '../action';
 
 function getSearchUrl(state : TinyMCEState) : string {
@@ -10,7 +10,7 @@ function getSearchUrl(state : TinyMCEState) : string {
 }
 export default function searchDelta(actions$ : Observable<Action>, state$ : Observable<TinyMCEState>) : Observable<Action> {
     return state$.sampledBy(actions$.filter(R.pipe(R.prop('type'), R.equals(SEARCH_INPUT))))
-        .flatMapLatest((state : TinyMCEState) : Observable<string, TypeError> => ajax(getSearchUrl(state), {
+        .flatMapLatest((state : TinyMCEState) : Observable<string, TypeError> => ajax$(getSearchUrl(state), {
             method: 'GET',
             headers: {
                 'X-WP-Nonce': state.globals.nonce
